携程促活撞库整体改造

See merge request pd_str/marketing!1091

+++ marketing-cs/src/main/java/com/br/marketing/mapper/XieChengCollidingDataLogMapperBase.java
@@ -14,7 +14,7 @@ public interface XieChengCollidingDataLogMapperBase {
 
     int insert(XieChengCollidingDataLog record);
 
-    int insertSelective(XieChengCollidingDataLog record);
+    Long insertSelective(XieChengCollidingDataLog record);
 
     List<XieChengCollidingDataLog> selectByExample(XieChengCollidingDataLogExample example);
